PHP 8.1 compatability when phpunit autoload file doesnt exit
oil attempts to load phpunits autoload file, ignoring failures
if it isn't able to load the file.  The code was using the @ operator
to supress errors in loading the file.

This change will check if the file exists as a gate in front of
loading it, which will prevent the error that is no longer
supressed in php 8.1